+++ src/main/java/com/onlyoffice/service/settings/SettingsValidationService.java
@@ -18,8 +18,6 @@
 
 package com.onlyoffice.service.settings;
 
-import com.onlyoffice.model.settings.HttpClientSettings;
-import com.onlyoffice.model.settings.security.Security;
 import com.onlyoffice.model.settings.validation.ValidationResult;
 
 
@@ -34,18 +32,6 @@ public interface SettingsValidationService {
     ValidationResult checkDocumentServer() throws Exception;
 
     /**
-     * Checks the Document Server by its URL.
-     *
-     * @param url The URL to the Document Server.
-     * @param httpClientSettings The settings for http client.
-     * @see HttpClientSettings
-     * @return The {@link ValidationResult} object.
-     * @throws Exception If the processing fails unexpectedly.
-     */
-    @Deprecated
-    ValidationResult checkDocumentServer(String url, HttpClientSettings httpClientSettings) throws Exception;
-
-    /**
      * Checks the Command Service.
      *
      * @return The {@link ValidationResult} object.
@@ -54,49 +40,10 @@ public interface SettingsValidationService {
     ValidationResult checkCommandService() throws Exception;
 
     /**
-     * Checks the Command Service by the Document Server URL.
-     *
-     * @param url The URL to the Document Server.
-     * @param security The security parameters.
-     * @param httpClientSettings The settings for http client.
-     * @see Security
-     * @return The {@link ValidationResult} object.
-     * @throws Exception If the processing fails unexpectedly.
-     */
-    @Deprecated
-    ValidationResult checkCommandService(String url, Security security, HttpClientSettings httpClientSettings)
-            throws Exception;
-
-    /**
-     * Checks the Convert Service.
-     *
-     * @return The {@link ValidationResult} object.
-     * @throws Exception If the processing fails unexpectedly.
-     */
-    @Deprecated
-    ValidationResult checkConvertService() throws Exception;
-
-    /**
      * Checks the Convert Service by the Document Server URL.
      *
      * @param productUrl The URL to the integration product.
      * @return The {@link ValidationResult} object.
      */
     ValidationResult checkConvertService(String productUrl);
-
-    /**
-     * Checks the Convert Service by the Document Server URL.
-     *
-     * @param url The URL to the Document Server.
-     * @param productInnerUrl The internal URL to the integration product.
-     * @param security The security parameters.
-     * @param httpClientSettings The settings for http client.
-     * @see Security
-     * @see HttpClientSettings
-     * @return The {@link ValidationResult} object.
-     * @throws Exception If the processing fails unexpectedly.
-     */
-    @Deprecated
-    ValidationResult checkConvertService(String url, String productInnerUrl, Security security,
-                                         HttpClientSettings httpClientSettings) throws Exception;
 }